Add tests for vendor constructor

+++ src/Models/Clothing.cs
@@ -1,6 +1,6 @@
 namespace cSharpSwapMeet
 {
-    public class Clothing(int itemID, string category = "Clothing", int condition = 0) : Item(itemID, category, condition)
+    public class Clothing(int itemID, string category = "Clothing", double condition = 0.0) : Item(itemID, category, condition)
     {
 
     }
