+++ src/Models/Decor.cs
@@ -1,6 +1,8 @@
+using System.ComponentModel;
+
 namespace cSharpSwapMeet
 {
-    public class Decor(int itemID, string category = "Decor", int condition = 0) : Item(itemID, category, condition)
+    public class Decor(int itemID, string category = "Decor", double condition = 0.0) : Item(itemID, category, condition)
     {
 
     }
