+++ src/Models/Electronics.cs
@@ -1,6 +1,6 @@
 namespace cSharpSwapMeet
 {
-    public class Electronics(int itemID, string category = "Electronics", int condition = 0) : Item(itemID, category, condition)
+    public class Electronics(int itemID, string category = "Electronics", double condition = 0.0) : Item(itemID, category, condition)
     {
 
     }
